requirements.txt
streamlit
openai
jinja2
faiss-cpu
sentence-transformers
